+++ src/main/java/roomescape/Controller/ReservationController.java
@@ -1,96 +1,43 @@
 package roomescape.Controller;
 
-import com.sun.jdi.LongValue;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import roomescape.Domain.Reservation;
+import roomescape.Service.ReservationService;
 
 import java.net.URI;
-import java.sql.PreparedStatement;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Objects;
-import java.util.concurrent.atomic.AtomicLong;
 
 @Controller
+@RequestMapping("/reservations")
 public class ReservationController {
-    private AtomicLong index = new AtomicLong(1);
+    // private AtomicLong index = new AtomicLong(1);
 
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
+    private final ReservationService reservationService;
 
-    public ReservationController(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
+    public ReservationController(ReservationService reservationService)
+    {
+        this.reservationService = reservationService;
     }
 
-    @GetMapping("/") // index로 이름을 바꾸어 처리했는데, @Controller 사용해서 수정
-    public String Home() {
-        return "home";
-    }
-
-    @GetMapping("/reservation")
-    public String Reservation() {
-        return "reservation";
-    }
-
-    @GetMapping("/reservations")
+    @GetMapping
     public ResponseEntity<List<Reservation>> Reservations() {
-        String sql = "select id, name, date, time from reservation";
-        List<Reservation> reservations = jdbcTemplate.query(sql, (resultSet, rowNum) ->
-        {
-            Reservation reservation = new Reservation(
-                    resultSet.getLong("id"),
-                    resultSet.getString("name"),
-                    resultSet.getString("date"),
-                    resultSet.getString("time")
-            );
-            return reservation;
-        });
+        List<Reservation> reservations = reservationService.getReservations();
         return ResponseEntity.ok().body(reservations);
     }
 
-    @PostMapping("/reservations")
+    @PostMapping
     public ResponseEntity<Reservation> createReservation(@RequestBody Reservation reservation) {
-        String sql = "insert into reservation (name, date, time) values (?, ?, ?)";
-
-        if (reservation == null || !reservation.isValid()) {
-            throw new IllegalArgumentException("누락된 사항이 있습니다. 확인해주세요.");
-        }
-
-        Long reservationId = insertWithKeyHolder(reservation);
+        Long reservationId = reservationService.createReservation(reservation);
         reservation.setId(reservationId);
 
         return ResponseEntity.created(URI.create("/reservations/" + reservation.getId())).body(reservation);
     }
 
-    @DeleteMapping("/reservations/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id)
-    {
-        String sql = "delete from reservation where id = ? ";
-        jdbcTemplate.update(sql, Long.valueOf(id));
-
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
+        reservationService.deleteReservation(id);
         return ResponseEntity.noContent().build();
-    }
-
-    public Long insertWithKeyHolder(Reservation reservation)
-    {
-        String sql = "insert into reservation (name, date, time) values (?, ?, ?)";
-
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
-            ps.setString(1, reservation.getName());
-            ps.setString(2, reservation.getDate());
-            ps.setString(3, reservation.getTime());
-            return ps;
-        }, keyHolder);
-
-        return keyHolder.getKey().longValue();
     }
 }
