+++ src/main/java/roomescape/Repository/ReservationRepository.java
@@ -1,14 +1,7 @@
 package roomescape.Repository;
 
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public class ReservationRepository {
-    private JdbcTemplate jdbcTemplate;
-
-    public ReservationRepository(JdbcTemplate jdbcTemplate)
-    {
-        this.jdbcTemplate = jdbcTemplate;
-    }
 }
